--- src/climate_learn/data_module/__init__.py
@@ -1,7 +1,0 @@
-from .data import *
-from .tasks import *
-from .constants import *
-from .download import *
-from .module import *
-from .itermodule import *
-from .nc2npz import convert_nc2npz
