--- src/climate_learn/data_module/data/__init__.py
@@ -1,4 +1,0 @@
-from .args import *
-from .data import *
-from .era5_module import *
-from .era5_iterdataset import *
