--- src/climate_learn/data_module/data/args/__init__.py
@@ -1,2 +1,0 @@
-from .data_args import *
-from .era5_args import *
