--- src/climate_learn/data_module/data/args/data_args.py
@@ -1,18 +1,0 @@
-from __future__ import annotations
-from abc import ABC
-from typing import Callable, Sequence, TYPE_CHECKING, Union
-
-if TYPE_CHECKING:
-    from climate_learn.data_module.data import Data
-    from climate_learn.data_module.module import DataModuleArgs
-
-
-class DataArgs(ABC):
-    _data_class: Union[Callable[..., Data], str] = "Data"
-
-    def __init__(self, variables: Sequence[str], split: str = "train") -> None:
-        self.variables: Sequence[str] = variables
-        self.split: str = split
-
-    def setup(self, data_module_args: DataModuleArgs) -> None:
-        pass
