+++ src/climate_learn/models/__init__.py
@@ -1,7 +1,7 @@
 from .components import *
 from .modules import *
-from climate_learn.data_module import IterDataModule, DataModule
-from climate_learn.data_module.tasks.args import ForecastingArgs
+from climate_learn.data import IterDataModule, DataModule
+from climate_learn.data.tasks.args import ForecastingArgs
 from climate_learn.utils.datetime import Hours
