+++ tests/data/climate_dataset/args/test_climate_dataset_args.py
@@ -1,0 +1,8 @@
+from climate_learn.data.climate_dataset.args import ClimateDatasetArgs
+
+
+class TestDataArgsInstantiation:
+    def test_initialization(self):
+        ClimateDatasetArgs(
+            variables=["random_variable_1", "random_variable_2"], split="train"
+        )
