--- tests/data_module/tasks/args/test_forecasting_args.py
@@ -1,14 +1,0 @@
-from climate_learn.data_module.tasks.args import ForecastingArgs
-from climate_learn.data_module.data.args import DataArgs
-
-
-class TestForecastingArgsInstantiation:
-    def test_initialization(self):
-        temp_data_args = DataArgs(
-            variables=["random_variable_1", "random_variable_2"], split="train"
-        )
-        ForecastingArgs(
-            temp_data_args,
-            in_vars=["random_variable_1"],
-            out_vars=["random_variable_2"],
-        )
